balance service test cases

+++ src/main/kotlin/com/example/splitwise/splitwise/service/BalanceService.kt
@@ -4,6 +4,6 @@ import com.example.splitwise.splitwise.module.Bill
 
 
 interface BalanceService {
-    fun getTotalBalance(userId: Long): Map<String, Long>
-    fun getIndividualBalance(userId: Long, respectUserId: Long): Map<String, Long>
+    fun getTotalBalance(userId: Long): Map<String, Double>
+    fun getIndividualBalance(userId: Long, respectUserId: Long): Map<String, Double>
 }
